fix: resolve build errors, Hilt DI, tests, and CI workflow tasks
- Add kotlin-parcelize plugin to version catalog and root/app build.gradle.kts
- Add missing dependencies: tvprovider, play-review
- Fix PlayerScreen.kt: missing Build import, brace balance, LazyColumn scope, engine smart-cast, setSecure cast
- Add missing PlayerViewModel methods: togglePlayPause, seekBackward, skipIntro/Outro, playNextEpisode, subtitle controls
- Fix MainActivity: OnBackInvokedDispatcher import, override onNewIntent
- Fix PlayerActivity: audioFocusRequest smart-cast
- Fix PlayerConfig: remove invalid setVideoSurfaceView block
- Fix FcmMessagingService: explicit Intent property assignment
- Fix RecommendationService: correct domain MediaItem mapping
- Fix MediaItem: add @IgnoredOnParcel for episodes list
- Inject WorkScheduler, CrashReporter, AppShortcuts, BatteryAwareManager in Application
- Fix ArcticThemeManager: remove duplicate getActiveSkin/setActiveSkin, use property syntax
- Fix ArcticFuseHomeScreen: add missing ThemeMode import
- Fix DataModule: provide Json, SubtitlePreferences, remove duplicate subtitle providers binding
- Fix KuroCloudModule: resolve OkHttpClient dependency cycle with @Named qualifiers
- Fix EncryptedPreferences: add @Inject constructor
- Fix KuroSubtitleEngine: use Set<SubtitleProvider> with JvmSuppressWildcards
- Fix test FavoritesRepositoryBridgeTest: update domain MediaItem/Profile fields, Result.failure
- Remove instrumentation-only PlayerConfigTest from unit tests
- Update CI workflows: ci.yml uses lintDebug, ci-cd.yml uses lintDebug, release.yml aligns signing secrets
- Add server package-lock.json for CI cache

+++ playback/src/main/java/com/kurostream/playback/kurovision/KuroVisionPipeline.kt
@@ -11,6 +11,7 @@ package com.kurostream.playback.kurovision
 import android.content.Context
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.asStateFlow
 import javax.inject.Inject
 import javax.inject.Singleton
 
@@ -46,7 +47,7 @@ class KuroVisionPipeline @Inject constructor(
 
         _frameStats.value = FrameStats(
             fps = currentFps,
-            frameCount = frameCount,
+            frameCount = frameCount.toLong(),
             dropCount = dropCount,
             lastProcessTimeUs = elapsedUs,
             isFallback = result.isPassthrough,
